requirements.txt
astroid==2.8.6
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
black==21.11b1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cryptography==36.0.0
fake-useragent==0.1.11
flake8==4.0.1
h11==0.12.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.6.0
loguru==0.5.3
lxml==4.6.4
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
outcome==1.1.0
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pylint==2.11.1
pyOpenSSL==21.0.0
regex==2021.11.10
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
toml==0.10.2
tomli==1.2.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.0
urllib3==1.26.7
wrapt==1.13.3
wsproto==1.0.0
